If given a container with variable heights, return the area
