+++ src/main/java/com/lperthel/datastructures/scratch.java
@@ -1,0 +1,5 @@
+package com.lperthel.datastructures;
+
+public class scratch {
+
+}
